Refactoring interfaces for Models
