+++ node-service/services/vulnerability-scan.service.ts
@@ -1,5 +1,6 @@
 // services/vulnerability-scan.service.ts
-import Scan, { IScan, IDependency, IVulnerability } from '../models/Scan';
+import Scan from '../models/Scan';
+import { IScan, IDependency, IVulnerability } from '../types/models';
 import osvService from './osv.service';
 
 interface VulnerabilityScanOptions {
